Added Feature Vectors for individual bots
Added a function which returns a dictionary containing the (name,
feature Vector) pair for each bot on the team.

As a note, total game time is constant (300 seconds)
All distances in the feature vector are normalized by the diagonal
distance of the entire map (the farthest conceivable distance possible)
but if this is incorrect feel free to change it.